+++ client/src/app/(withauth)/products/datalake/dataset/page.tsx
@@ -1,6 +1,6 @@
 "use client"
-import { Badge, Container, Row } from "react-bootstrap"
-import { Fragment } from "react"
+import { Badge, Col, Container, Row } from "react-bootstrap"
+import { Fragment, useCallback } from "react"
 import Loading from "@/components/loading-component"
 import Show from "@/components/show-component"
 import { endPoints } from "@/constants/api-endpoints"
@@ -8,7 +8,7 @@ import useFetch from "@/hooks/use-fetch"
 import HTTPMethods from "@/constants/http-methods"
 import Error from "@/components/error-component"
 import { useSearchParams } from "next/navigation"
-import { CubeIcon } from "@radix-ui/react-icons"
+import { BookmarkIcon, CubeIcon } from "@radix-ui/react-icons"
 import { ProductCardInterface } from "@/types/Types"
 import ProductCard from "@/components/productcard-component"
 import Hero from "@/components/hero-component"
@@ -28,7 +28,12 @@ export default function Page() {
       footerText: `${dataset.description.slice(0, 110)}...`,
       redirectUri: `/products/datalake/dataset?datasetId=${dataset._id}`
     }
-    return <ProductCard key={dataset._id} productCardProps={productCardProps} />
+
+    return (
+      <Col xs={12} sm={6} md={6} lg={4} xl={3} className="mb-4">
+        <ProductCard key={dataset._id} productCardProps={productCardProps} />
+      </Col>
+    )
   })
 
   const datasetTagsToDisplay = dataset?.data?.description?.split(" ").slice(0, 30).map((item: string) => {
@@ -36,6 +41,21 @@ export default function Page() {
       return <Badge bg="light" className="mt-2 me-2 top-0 end-0 ps-3 pe-3 p-2" key={Math.random().toString()}>{item}</Badge>
     }
   })
+
+  const datasetQuality = useCallback(() => {
+    const rating = dataset?.data?.rating
+    if (rating > 4.5) {
+      return <Badge className="gold-badge mt-2 me-2 top-0 end-0 ps-3 pe-3 p-2" key={Math.random().toString()}><BookmarkIcon className="icon-left" />Gold</Badge>
+    }
+
+    else if (rating > 4.0 && rating < 4.5) {
+      return <Badge className="silver-badge mt-2 me-2 top-0 end-0 ps-3 pe-3 p-2" key={Math.random().toString()}><BookmarkIcon className="icon-left" />Silver</Badge>
+    }
+
+    else {
+      return <Badge className="bronze-badge mt-2 me-2 top-0 end-0 ps-3 pe-3 p-2" key={Math.random().toString()}><BookmarkIcon className="icon-left" />Bronze</Badge >
+    }
+  }, [dataset?.data?.rating])
 
   return (
     <Fragment>
@@ -45,6 +65,7 @@ export default function Page() {
             <Hero>
               <p className="branding text-capitalize">{dataset?.data?.name}</p>
               <p className="lead">{dataset?.data?.category}</p>
+              {datasetQuality()}
               <p className="muted-text mt-3">{dataset?.data?.description}</p>
               <div className="mb-4">{datasetTagsToDisplay}</div>
               <SensitiveInfoPanel credentialIcon={<CubeIcon />} credentialName="Dataset ID" credentialValue={datasetId ?? ""} />