+++ client/src/app/(withauth)/products/layout.tsx
@@ -1,12 +1,21 @@
 "use client"
 import Show from "@/components/show-component"
 import { GlobalContext } from "@/context/globalstate.provider"
-import { BookmarkIcon, CubeIcon } from "@radix-ui/react-icons"
+import { BookmarkIcon, CubeIcon, HomeIcon } from "@radix-ui/react-icons"
 import Link from "next/link"
+import { usePathname, useRouter } from "next/navigation"
 import { Fragment, ReactNode, useContext } from "react"
 
 export default function ProductLayout({ children }: { children: ReactNode }) {
   const [{ userState }] = useContext(GlobalContext)
+  const pathname = usePathname()
+  const router = useRouter()
+  const urlParts = pathname.split("/")
+  const productsIndex = urlParts.indexOf('products')
+  let productName = ""
+  if (productsIndex !== -1 && productsIndex + 1 < urlParts.length) {
+    productName = urlParts[productsIndex + 1]
+  }
 
   return (
     <Fragment>