+++ client/src/app/(withauth)/products/nftstudio/page.tsx
@@ -5,7 +5,7 @@ import { endPoints } from "@/constants/api-endpoints"
 import Web3 from "web3"
 import Link from "next/link"
 import { Fragment, useCallback, useContext, useEffect, useState } from "react"
-import { Badge, Container, Row } from "react-bootstrap"
+import { Badge, Col, Container, Row } from "react-bootstrap"
 import { toast } from "react-hot-toast"
 import { GlobalContext } from "@/context/globalstate.provider"
 import { nftABI } from "@/bin/nft-abi"
@@ -64,7 +64,9 @@ export default function Page() {
       }
 
       return (
-        <ProductCard key={nft.id} productCardProps={productCardProps} />
+        <Col xs={12} sm={6} md={6} lg={4} xl={3} className="mb-4">
+          <ProductCard key={nft.id} productCardProps={productCardProps} />
+        </Col>
       )
     })