+++ client/src/components/header-component.tsx
@@ -4,11 +4,12 @@ import { Container, Navbar, Nav } from "react-bootstrap"
 import Show from "./show-component"
 import { usePathname, useSearchParams } from "next/navigation"
 import Link from "next/link"
-import { HamburgerMenuIcon, TextAlignLeftIcon } from "@radix-ui/react-icons"
+import { BellIcon, GearIcon, TextAlignRightIcon } from "@radix-ui/react-icons"
 import { uiConstants } from "@/constants/global-constants"
 import debounce from "lodash.debounce"
 import { GlobalContext } from "@/context/globalstate.provider"
 import { useRouter } from "next/navigation"
+import { uiHost } from "@/constants/api-endpoints"
 
 interface HeaderProps {
   isAuthorized: boolean,
@@ -62,22 +63,27 @@ export default function Header({ isAuthorized }: HeaderProps) {
         <Navbar variant="light" expand="lg" fixed="top" className="pt-3 pb-3">
           <Container>
             <Link href="/dashboard">
-              <Navbar.Brand>{uiConstants.brandName}</Navbar.Brand>
+              <Navbar.Brand>
+                <img alt="" src={`${uiHost}/favicon.ico`} width="30" height="30" className="d-inline-block align-top" />{" "}
+                {uiConstants.brandName}
+              </Navbar.Brand>
             </Link>
             <Navbar.Toggle>
-              <HamburgerMenuIcon className="icon-nav-toggle" />
+              <TextAlignRightIcon className="icon-nav-toggle" />
             </Navbar.Toggle>
             <Navbar.Collapse>
-              <Show when={searchEnabledPathNames.includes(pathname)}>
-                <Nav className="ms-auto">
-                  <input ref={searchRef} placeholder="Press (Alt + Q) or click here to search" type="text" className="header-search" onChange={debouncedChangeHandler} />
-                </Nav>
-              </Show>
-              <Nav className="ms-auto">
+              <Nav className="mr-auto">
                 <Nav.Item><Link href="/workspace">Workspace</Link></Nav.Item>
                 <Nav.Item><Link href="/subscription">Subscription</Link></Nav.Item>
                 <Nav.Item><Link target="_blank" passHref rel="noopener noreferrer" href={uiConstants.mumbaiFaucetUri}>Fund Wallet</Link></Nav.Item>
                 <Nav.Item className="btn-user-link"><Link href="/account">Account</Link></Nav.Item>
+                <Nav.Item className="btn-user-link"><Link href="/activities">Activities</Link></Nav.Item>
+              </Nav>
+              <Nav className="ms-auto">
+                <Show when={searchEnabledPathNames.includes(pathname)}>
+                  <input ref={searchRef} placeholder="Press (Alt + Q) or click here to search" type="text" className="header-search" onChange={debouncedChangeHandler} />
+                </Show>
+                <BellIcon className="icon-navbar" onClick={() => router.push("/activities")} />
                 <button className="btn-user" onClick={(): void => router.push("/account")}>{userInitial}</button>
               </Nav>
             </Navbar.Collapse>
@@ -88,10 +94,13 @@ export default function Header({ isAuthorized }: HeaderProps) {
         <Navbar variant="light" expand="lg" fixed="top" className="pt-3 pb-3">
           <Container>
             <Link href="/">
-              <Navbar.Brand>{uiConstants.brandName}</Navbar.Brand>
+              <Navbar.Brand>
+                <img alt="" src={`${uiHost}/favicon.ico`} width="30" height="30" className="d-inline-block align-top" />{" "}
+                {uiConstants.brandName}
+              </Navbar.Brand>
             </Link>
             <Navbar.Toggle>
-              <HamburgerMenuIcon className="icon-nav-toggle" />
+              <TextAlignRightIcon className="icon-nav-toggle" />
             </Navbar.Toggle>
             <Navbar.Collapse>
               <Nav className="ms-auto">