+++ src/config/activity.config.ts
@@ -1,0 +1,14 @@
+type ActivityConfig = {
+  [key: string]: string
+}
+
+export const activityConfig: ActivityConfig = {
+  "/api/workspace/create": "Create Workspace",
+  "/api/workspace/switch": "Switch Workspace",
+  "/api/subscription/activatetrial": "Activate Trial Plan",
+  "/api/products/fabric/createdb": "Create Fabric DB",
+  "/api/products/fabric/deletedb": "Delete Fabric DB",
+  "/api/products/insights/createproject": "Create Insights Project",
+  "/api/products/insights/deleteproject": "Delete Insights Project",
+  "/api/products/datalake/viewdataset": "Browse Datalake Dataset",
+}